Add normalizePivot to polyfill.ts and rename to nodeWorkarounds.ts
